requirements.txt
pandas>=2.0
numpy
plotly>=5.0
click>=8.0
watchdog>=3.0
fastapi>=0.100
uvicorn>=0.20
python-dotenv>=1.0
slowapi>=0.1.9
streamlit>=1.30
httpx>=0.24
pytest>=7.0
black>=24.0
flake8>=7.0
isort>=5.13
pre-commit>=3.0
